This is a collection of Projects
This is a collection of projects done in parallel of the Colt Steele WebDev Course.
The Sample Page is a project I did to learn how to make custom styling.
My first custom page is a project made to go further into the styling and also combine. HTML, CSS, and JavaScript.